Rename harborList __mocks__ to __fixtures__, move tests to __tests__
